qwen2.5

+++ transformer_lens/components/rms_norm.py
@@ -38,10 +38,15 @@ class RMSNorm(nn.Module):
     ) -> Float[torch.Tensor, "batch pos length"]:
         if self.cfg.dtype not in [torch.float32, torch.float64]:
             x = x.to(torch.float32)
+        # scale: Float[torch.Tensor, "batch pos 1"] = self.hook_scale(
+        #     (x.pow(2).mean(-1, keepdim=True) + self.eps).sqrt()
+        # )
+        # x = self.hook_normalized(x / scale).to(self.cfg.dtype)  # [batch, pos, length]
+
         scale: Float[torch.Tensor, "batch pos 1"] = self.hook_scale(
-            (x.pow(2).mean(-1, keepdim=True) + self.eps).sqrt()
+            (x.pow(2).mean(-1, keepdim=True) + self.eps).sqrt().to(self.cfg.dtype)
         )
-        x = self.hook_normalized(x / scale).to(self.cfg.dtype)  # [batch, pos, length]
+        x = self.hook_normalized((x / scale).to(self.cfg.dtype))  # [batch, pos, length] #hsc
 
         if x.device != self.w.device:
             self.to(x.device)
